Send the goblin._init command for all Xcraft nodes
Previously, only the main goblin was used to init the core goblin. It
was completly wrong. Now, all node have its own core goblin initialized.

Note that this call must not use the warehouse because the warehouse is
uniq and can exists only in one process. Then it's possible that the
warehouse is not ready when a goblin._init call is executing.