Remove globalThis which is no longer necessary
With Object.assign the reference is no longer change and the require is
done only one time in the process. In this case we are just overriding
in the existing object.

A _isMinimalConfig boolean is stored in the xHost object in order to
have a way to detect if the use of xHost is before or after the load.